// app-servicioexpress/src/app/models/ClientServiceModel.ts
export interface CustomerServiceModel {
    id: number;
    idService: number;
    idStatusService: number;
    name: string;
    phone: string;
    curp: string;
    nss: string;
    urlSemanasCotizadas: string;
    urlVigenciaDerechos: string;
    statusService: string;
    asesor: string;
    fechaAlta: string;
    urlTicketPago: string;
}